auth_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
drf-spectacular==0.26.5

# Optionnel - pour PostgreSQL (nécessite Visual C++ Build Tools)
# psycopg2-binary==2.9.9

# Optionnel - pour Redis cache
# django-redis==5.4.0
# python-decouple==3.8

